Pull out ChatManagerGrid as a component and tidy
