+++ admin_app/src/app/integrations/components/ChatManagerModals.tsx
@@ -1,4 +1,4 @@
-import { appColors, appStyles, sizes } from "@/utils";
+import { appColors, sizes } from "@/utils";
 import {
   Box,
   Button,
